pyproject.toml
[tool.poetry]
name = "product-groups"
version = "0.1.0"
description = "Анализ чеков и товарных групп"
authors = ["ilyamikheev <r1sephoenix@yandex.ru>"]
packages = [
    { include = "product_groups", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pandas = "^2.1.4"
scikit-learn = "^1.3.2"
jinja2 = "^3.1.2"
mlxtend = "^0.23.1"
tqdm = "^4.66.1"
typer = "^0.9.0"
python-dotenv = "^1.0.0"
apache-airflow = "^2.7.0"
clickhouse-driver = "^0.2.6"
boto3 = "^1.28.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
mypy = "^1.0.0"
black = "^23.0"

[tool.poetry.scripts]
store-calc = "cli.commands:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"